// packages/core/src/interfaces/IApplication.ts
export enum ApplicationEvent {
  "@INIT" = '@application-init',
  BEFORE_START = 'before-start',
  ON_START = 'on-start',
  AFTER_START = 'after-start',
  BEFORE_CONFIG = 'before-config',
  AFTER_CONFIG = 'after-config',
}